refactor(macos): extract app update coordinator
